codebase: add title cache subinterface.
Used by the existing title interface to load control data off the SD card using a custom binary format to avoid cluttering the SD card filesystem.

This title cache subinterface is designed to work with RAM-based operations as much as possible, keeping SD card I/O operations to a minimum. At the time of this writing, the title cache file is saved to "sdmc:/switch/title_cache.bin" (this may change in the future).

This was done in the hopes of drastically reducing load times under HOS 20.0.0 and above.

Other changes include:

* poc: display a "please wait" message at startup.

* title: update code to use the new title cache subinterface.
* title: update relevant functions to make them retrieve control data from the title cache subinterface first, and fallback to other methods if it's not available there.
* title: update code to manually flush the title cache file at very specific points.
* title: update code to completely avoid using ns calls to retrieve control data under HOS 20.0.0+ unless it's absolutely necessary.